refactor upgrade v0.2.2
