+++ app/upgrade_v0_2_1.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	upgradetypes "cosmossdk.io/x/upgrade/types"
-	"github.com/Galactica-corp/galactica/app/upgrades/v0_2_1"
+	"github.com/Galactica-corp/galactica/app/upgrades/v0_2_2"
 
 	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -12,33 +12,34 @@ import (
 )
 
 func (app *App) applyUpgrade_v0_2_2() {
-	logger := app.Logger().With("upgrade", v0_2_1.UpgradeName)
+	planName := v0_2_2.UpgradeName
 
-	planName := v0_2_1.UpgradeName
-	app.UpgradeKeeper.SetUpgradeHandler(planName, app.upgradeHandler_v0_2_1())
+	logger := app.Logger().With("upgrade", planName)
+
+	app.UpgradeKeeper.SetUpgradeHandler(planName, app.upgradeHandler_v0_2_2())
 	
 	ctx := app.NewContext(true).WithBlockHeader(tmproto.Header{Height: app.LastBlockHeight()}).
 		WithMultiStore(app.CommitMultiStore())
 
-	doneHeight, err := app.UpgradeKeeper.GetDoneHeight(ctx, v0_2_1.Plan.Name)
+	doneHeight, err := app.UpgradeKeeper.GetDoneHeight(ctx, planName)
 	if err != nil {
 		logger.Error("Error with GetDoneHeight", "error", err)
 		return
 	}
 
 	if doneHeight != 0 {
-		logger.Info("upgrade v0.2.1 done")
+		logger.Info("upgrade v0.2.2 done")
 		return
 	}
 
-	logger.Info("Schedule upgrade plan", "info", v0_2_1.Plan.Info)
+	logger.Info("Schedule upgrade plan", "name", planName)
 
-	if err := app.UpgradeKeeper.ScheduleUpgrade(ctx, v0_2_1.Plan); err != nil {
+	if err := app.UpgradeKeeper.ScheduleUpgrade(ctx, v0_2_2.Plan); err != nil {
 		panic(err)
 	}
 }
 
-func (app *App) upgradeHandler_v0_2_1() func(ctx context.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+func (app *App) upgradeHandler_v0_2_2() func(ctx context.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 	return func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 		logger := sdk.UnwrapSDKContext(ctx).Logger()
